Insert step for setting predictable iface names.
In router mode, where we have two WLAN interfaces to run, we want to
assign "wlan0" name for the builtin interface and "wlan1" for any second
interfaces.

To ensure, that this really happens, we introduce a `udev` rule which is
applied at bootup and assigns the names based on the bus type used by
the interfaces.